Add workflow package and integrate with web app
Introduced the @jitaspace/workflow package with initial configuration, lint, and TypeScript setup. Updated the web app to use the new workflow package in next.config.ts, added it to dependencies, and included its TypeScript plugin. Migrated next.config from .mjs to .ts and updated relevant configuration to support workflow integration.
