--- apps/web/next.config.mjs
@@ -1,102 +1,0 @@
-import bundleAnalyzer from "@next/bundle-analyzer";
-import { withSentryConfig } from "@sentry/nextjs";
-import { createJiti } from "jiti";
-
-const jiti = createJiti(import.meta.url);
-/**
- * Run `build` or `dev` with `SKIP_ENV_VALIDATION` to skip env validation.
- * This is especially useful for Docker builds and Linting.
- */
-!process.env.SKIP_ENV_VALIDATION && (await jiti.import("./env"));
-
-/** @type {import("next").NextConfig} */
-const config = {
-  reactStrictMode: true,
-
-  /** Enables hot reloading for local packages without a build step */
-  transpilePackages: [
-    "@jitaspace/auth",
-    "@jitaspace/db",
-    "@jitaspace/esi-metadata",
-    "@jitaspace/esi-client",
-    "@jitaspace/esi-meta-client",
-    "@jitaspace/eve-icons",
-    "@jitaspace/eve-scrape",
-    "@jitaspace/hooks",
-    "@jitaspace/kv",
-    "@jitaspace/sde-client",
-    "@jitaspace/tiptap-eve",
-    "@jitaspace/ui",
-    "@jitaspace/utils",
-  ],
-
-  /** We already do typechecking as separate tasks in CI */
-  typescript: { ignoreBuildErrors: !!process.env.CI },
-
-  /** Allow images from CCP CDN */
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "web.ccpgamescdn.com",
-      },
-      {
-        protocol: "https",
-        hostname: "images.evetech.net",
-      },
-    ],
-  },
-
-  rewrites: async () => [
-    {
-      source: "/analytics/:match*",
-      destination: "https://analytics.umami.is/:match*", // Proxy to Umami
-    },
-  ],
-
-  env: {
-    NEXT_PUBLIC_MODIFIED_DATE: new Date().toISOString(),
-  },
-};
-
-const withBundleAnalyzer = bundleAnalyzer({
-  enabled: process.env.ANALYZE === "true",
-});
-
-export default withSentryConfig(withBundleAnalyzer(config), {
-  // For all available options, see:
-  // https://www.npmjs.com/package/@sentry/webpack-plugin#options
-
-  org: "jitaspace",
-
-  project: "jitaspace",
-
-  // Only print logs for uploading source maps in CI
-  silent: !process.env.CI,
-
-  // For all available options, see:
-  // https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/
-
-  // Upload a larger set of source maps for prettier stack traces (increases build time)
-  widenClientFileUpload: false,
-
-  // Route browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers.
-  // This can increase your server load as well as your hosting bill.
-  // Note: Check that the configured route will not match with your Next.js middleware, otherwise reporting of client-
-  // side errors will fail.
-  tunnelRoute: "/monitoring",
-
-  webpack: {
-    // Enables automatic instrumentation of Vercel Cron Monitors. (Does not yet work with App Router route handlers.)
-    // See the following for more information:
-    // https://docs.sentry.io/product/crons/
-    // https://vercel.com/docs/cron-jobs
-    automaticVercelMonitors: true,
-
-    // Tree-shaking options for reducing bundle size
-    treeshake: {
-      // Automatically tree-shake Sentry logger statements to reduce bundle size
-      removeDebugLogging: true,
-    },
-  },
-});
