+++ packages/workflow/index.ts
@@ -1,0 +1,3 @@
+import { withWorkflow } from "workflow/next";
+
+export { withWorkflow };
