+++ packages/workflow/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@jitaspace/tsconfig/base.json",
+  "compilerOptions": {
+    "lib": ["es2022", "dom", "dom.iterable"],
+    "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
+  },
+  "include": ["."],
+  "exclude": ["node_modules"]
+}
